requirement.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
arviz==0.14.0
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
biopython==1.79
bleach==4.1.0
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.11.8
cffi==1.15.0
cftime==1.6.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
cmdstanpy==0.9.5
colorama==0.4.4
contourpy==1.0.5
convertdate==2.4.0
cryptography==39.0.1
cvxopt==1.3.0
cycler==0.10.0
Cython==0.29.33
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.3
ephem==4.1.4
executing==1.2.0
prophet==1.1
fitter==1.5.2
flatbuffers==1.12
flit_core==3.6.0
fonttools==4.25.0
frozenlist==1.3.3
fst-pso==1.8.1
future==0.18.3
FuzzyTM==1.0.0
gast==0.4.0
gensim==4.3.0
google-auth==2.6.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.15
grpcio==1.38.1
h5py==2.10.0
hijri-converter==2.2.4
holidays==0.21
hyperopt==0.2.5
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.4.0
intervaltree==3.1.0
ipdb==0.13.13
ipython==8.11.0
ipython-genutils==0.2.0
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.0.3
joblib==1.1.1
jsonpickle==3.0.1
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
miniful==0.0.6
mistune==0.8.4
mpmath==1.2.1
multidict==6.0.2
munkres==1.1.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
nltk==3.7
numexpr==2.8.4
numpy==1.19.5
oauthlib==3.2.1
opt-einsum==3.3.0
opyenxes==0.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
pm4py==2.2.32
pooch==1.4.0
powerlaw==1.5
prometheus-client==0.12.0
prompt-toolkit==3.0.38
protobuf==3.20.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydotplus==2.0.2
pyFUME==0.2.25
Pygments==2.10.0
PyJWT==2.4.0
PyMeeus==0.5.12
pyOpenSSL==23.0.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2022.7.1
pyvis==0.3.2
PyYAML==6.0
pyzmq==22.3.0
regex==2022.7.9
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.1.2
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.6.3
setuptools-git==1.2
setuptools-scm==7.1.0
simpful==2.10.0
six==1.16.0
smart-open==5.2.1
sortedcontainers==2.4.0
stack-data==0.6.2
StringDist==1.0.9
sympy==1.7.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.3
tensorflow-estimator==2.5.0
termcolor==2.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wheel==0.38.4
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
xarray==2022.11.0
xarray-einstats==0.3.0
xlrd==1.2.0
yarl==1.8.1
zipp==3.6.0